Address review iteration 2: whatsnew entry + Py_ssize_t
- Add a 'faulthandler' entry to Doc/whatsnew/3.15.rst (per
  @ZeroIntensity).
- Switch the max_threads parameters/fields and the nthreads loop
  counter in _Py_DumpTracebackThreads to Py_ssize_t (per @vstinner).

+++ Include/internal/pycore_faulthandler.h
@@ -57,7 +57,7 @@ struct _faulthandler_runtime_state {
         void *exc_handler;
 #endif
         int c_stack;
-        unsigned int max_threads;
+        Py_ssize_t max_threads;
     } fatal_error;
 
     struct {
@@ -69,7 +69,7 @@ struct _faulthandler_runtime_state {
         int exit;
         char *header;
         size_t header_len;
-        unsigned int max_threads;
+        Py_ssize_t max_threads;
         /* The main thread always holds this lock. It is only released when
            faulthandler_thread() is interrupted before this thread exits, or at
            Python exit. */
